% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinWData.R
\name{binWData}
\alias{binWData}
\title{binWData}
\usage{
binWData(data, base = 4, min_size = 0.1)
}
\arguments{
\item{dataframe}{}
}
\value{
dataframe
}
\description{
  Logarithmic binning of asymptotic sizes using base=base. 
  The lowest bin contains indviduals with asymptotic size in the range
  \{min_size/base;min_size/base\}
}
\examples{
....
....
....

}
